refactor: #406
좌석 상태 조회 시 section별 seat ID 조회 후 상태 조회하도록 변경
