+++ ticketing/src/main/java/com/goti/ticketing/seat/service/domain/SeatService.java
@@ -12,5 +12,7 @@ public interface SeatService {
 
 	List<SeatResponse> get(UUID sectionId, UUID gameId, UUID userId);
 
+	List<UUID> getSeatIdsBySectionId(UUID sectionId);
+
 	List<SeatEntity> getByStadiumId(UUID stadiumId);
 }
