+++ ticketing/src/main/java/com/goti/ticketing/seat/service/domain/SeatStatusService.java
@@ -8,11 +8,8 @@ import java.util.UUID;
 import com.goti.ticketing.domain.entity.game.GameScheduleEntity;
 import com.goti.ticketing.domain.entity.seat.SeatEntity;
 import com.goti.ticketing.domain.entity.seat.SeatStatusEntity;
-import com.goti.ticketing.seat.dto.response.GameSeatStatusResponse;
 
 public interface SeatStatusService {
-	List<GameSeatStatusResponse> get(UUID gameId, UUID sectionId, UUID userId);
-
 	Map<UUID, SeatStatusEntity> getByGameIdAndSeatIds(UUID gameId, List<UUID> seatIds);
 
 	Set<UUID> getSeatIdsByGameId(UUID gameId);
